Lets play with files!
So the code so far can commit files to the repository, storing some basic
metadata along with it.

Additionally, 'requests' can be registered, and files accepted for a particular request.
When a file comes in for a request, that row records the file id associated with it. At
that point it's ready for a periodic task on the behalf of the bot to tell people where
the file is and possibly remove the request row.

Not sure about that last bit, I'm not sure I want the log to be the authoritive link
to the file...



git-svn-id: svn+ssh://romana.hawkins.id.au/home/justin/repos/CollabIRCate/trunk@161 8d8e6d39-c747-0410-aa99-ea999ebef502