Allow for overriding configuration variables with ENV variables.
I don't like this much. It's different to the way the internal-to-the-app
configuration overrides work, but after discussion with mst on the #catalyst
list it seems there is currently not a lot of choice.

But for now it will help a lot with testing.

Also fixed a bug with the bot always connecting to localhost.



git-svn-id: svn+ssh://romana.hawkins.id.au/home/justin/repos/CollabIRCate/trunk@164 8d8e6d39-c747-0410-aa99-ea999ebef502