This is pretty hideous yet.
Takes a MIME message on stdin, if it find a hash in the To: field (via
plus addressing), any non text/plain attachments will be added to the
CollabIRCate database.

Still lots to do, including a horendous amount of test harness work
to properly test this. I think I'll just hack it together first :-)



git-svn-id: svn+ssh://romana.hawkins.id.au/home/justin/repos/CollabIRCate/trunk@174 8d8e6d39-c747-0410-aa99-ea999ebef502