Add testdata (raw emails with mime attachments).
Link each file with the request that created it, and created
a flag which shows if we've announced it in the channel log
yet. This will let me add code to the bot to periodically scan
for unlogged files and announce them to the channel. And we'll all
be like 'thanks peoplebot!'.



git-svn-id: svn+ssh://romana.hawkins.id.au/home/justin/repos/CollabIRCate/trunk@175 8d8e6d39-c747-0410-aa99-ea999ebef502