Start and stop the irc server and bot as part of the test.
Start and stop the server before running the tests, to allow for
automated tests which use irc_test.pl or other irc clients.

Also an ugly hack to allow us to pause the harness in the middle
to run other manual tests. This is triggered by specifying anything
at all on the command line.