Improve interactions with the bot.
Paving the way for better plugins, they can register a method to be called
when they see any message, one for when they are addressed directly, and
another for periodic actions.

Still need to filter all this up to the top level.

In particular not sure what the return values should be. Arguments to
an $irc->yield?